requirements.txt
asgiref==3.6.0
astroid==2.15.6
async-timeout==4.0.3
attrs==22.2.0
autobahn==23.6.2
Automat==22.10.0
bandit==1.7.5
black==23.7.0
boltons==21.0.0
bracex==2.3.post1
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.5
click-option-group==0.5.6
colorama==0.4.6
constantly==15.1.0
coverage==7.2.7
crashtest==0.4.1
cryptography==41.0.1
daphne==4.0.0
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
Django==4.2.5
django-allauth==0.54.0
django-cors-headers==4.2.0
django-csp==3.7
djangorestframework==3.14.0
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.3.0
dodgy==0.2.1
dulwich==0.21.3
exceptiongroup==1.1.2
face==22.0.0
filelock==3.10.7
flake8==5.0.4
flake8-polyfill==1.0.2
gitdb==4.0.10
GitPython==3.1.32
glom==22.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.1.0
incremental==22.10.0
iniconfig==2.0.0
installer==0.7.0
isort==5.12.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
lazy-object-proxy==1.9.0
lockfile==0.12.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
msgpack==1.0.5
mypy==1.4.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.1
pbr==5.11.1
peewee==3.16.2
pep8-naming==0.10.0
pexpect==4.8.0
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.2.0
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.0
prospector==1.10.2
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.9.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.5.0
Pygments==2.15.1
PyJWT==2.8.0
pylint==2.17.4
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pyOpenSSL==23.2.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-mock==3.11.1
python-lsp-jsonrpc==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==2.13.7
redis==5.0.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
requirements-detector==1.2.2
rich==13.4.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
semgrep==1.34.1
semver==3.0.1
service-identity==23.1.0
setoptconf-tmp==0.3.1
shellingham==1.5.0.post1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
stevedore==5.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
trove-classifiers==2023.3.9
Twisted==23.8.0
txaio==23.1.1
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.7.1
ua-parser==0.18.0
ujson==5.8.0
urllib3==1.26.15
virtualenv==20.21.0
wcmatch==8.4.1
webencodings==0.5.1
whitenoise==6.5.0
wrapt==1.15.0
xattr==0.10.1
zipp==3.15.0
zope.interface==6.1
